% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_L2_distance}
\alias{compute_L2_distance}
\title{Compute L2 Distance}
\usage{
compute_L2_distance(vec1, vec2)
}
\arguments{
\item{vec1}{First vector.}

\item{vec2}{Second vector.}
}
\value{
The L2 distance between vec1 and vec2.
}
\description{
Computes the L2 distance between two vectors.
}
\keyword{internal}
