% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gcd}
\alias{gcd}
\title{Greatest Common Divisor}
\usage{
gcd(n, m)
}
\arguments{
\item{n}{First integer.}

\item{m}{Second integer.}
}
\value{
The greatest common divisor of n and m.
}
\description{
Computes the greatest common divisor of two integers.
}
\keyword{internal}
