% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grem.R
\name{grem}
\alias{grem}
\title{Global Regression for Empirical Measures (GREM)}
\usage{
grem(y = NULL, x = NULL, xOut = NULL, optns = list())
}
\arguments{
\item{y}{A list of empirical distributions (each element is a vector of observations).}

\item{x}{A matrix, data frame, or vector of predictors.}

\item{xOut}{Optional matrix of out-of-sample predictor values for prediction.}

\item{optns}{Optional list of control parameters (e.g., upper, lower bounds).}
}
\value{
An object of class "rem" containing:
  \item{qf}{Fitted quantile functions (n x M matrix)}
  \item{qfSupp}{Support points for quantile functions}
  \item{qp}{Predicted quantile functions at xOut (if provided)}
  \item{qpSupp}{Support points for predictions}
  \item{RSquare}{R-squared value}
  \item{adjRSquare}{Adjusted R-squared value}
  \item{residuals}{Vector of residuals}
  \item{y}{Input empirical distributions}
  \item{x}{Input predictors}
  \item{xOut}{Out-of-sample predictors}
  \item{optns}{Options used}
}
\description{
Performs global Frechet regression using the Wasserstein metric for
distributional responses.
}
\examples{
\dontrun{
# Generate some data
n <- 100
x <- runif(n)
y <- lapply(1:n, function(i) rnorm(50, mean = x[i], sd = 0.1))
fit <- grem(y, x, xOut = c(0.3, 0.5, 0.7))
}
}
