% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lcm}
\alias{lcm}
\title{Least Common Multiple}
\usage{
lcm(n, m)
}
\arguments{
\item{n}{First integer.}

\item{m}{Second integer.}
}
\value{
The least common multiple of n and m.
}
\description{
Computes the least common multiple of two integers.
}
\keyword{internal}
