% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{simulate_data_setting2}
\alias{simulate_data_setting2}
\title{Simulate Data - Setting 2}
\usage{
simulate_data_setting2(n_t, n_vec, M, K = 5)
}
\arguments{
\item{n_t}{Target sample size.}

\item{n_vec}{Vector of source sample sizes (length K).}

\item{M}{Grid size for quantile functions.}

\item{K}{Number of source datasets (default 5).}
}
\value{
A list with components:
  \item{X_t}{Target predictors (n_t x 1)}
  \item{Y_t}{Target quantile functions (n_t x M)}
  \item{source_data}{List of K source datasets, each containing X_s, Y_s, and d_k}
}
\description{
Generates target and source data according to Setting 2 specifications from
the Wasserstein Transfer Learning paper.
}
\examples{
\dontrun{
data <- simulate_data_setting2(n_t = 200, n_vec = c(100, 200, 300, 400, 500), M = 100, K = 5)
}
}
