% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_matrix}
\alias{to_matrix}
\title{Convert to Matrix}
\usage{
to_matrix(x)
}
\arguments{
\item{x}{A vector, data frame, or matrix.}
}
\value{
A matrix representation of x.
}
\description{
Helper function to convert vectors or data frames to matrices.
}
\keyword{internal}
